优化CheckIn源码结构
